+++ cloudsim-plus-examples/src/main/java/org/cloudsimplus/examples/StorageAreaNetworkExample1.java
@@ -146,7 +146,7 @@ import java.util.List;
             sanList.add(san);
         }
 
-        dc.setStorageList(sanList);
+        dc.getDatacenterStorage().setStorageList(sanList);
     }
 
     /**
